# doc/openapi.yaml
openapi: 3.1.0
info:
  title: Social Network API
  version: 0.1.0
paths:
  /post:
    post:
      summary: Создание поста
      operationId: create_post_post_post
      parameters:
        - name: session-key
          in: header
          required: true
          schema:
            type: string
            format: uuid
            title: Session-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { }
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      summary: Обновление контента поста
      operationId: update_post_post_put
      parameters:
        - name: session-key
          in: header
          required: true
          schema:
            type: string
            format: uuid
            title: Session-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { }
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  '/post/{post_id}':
    delete:
      summary: Удаление поста
      operationId: delete_post_post__post_id__delete
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
            title: Post Id
        - name: session-key
          in: header
          required: true
          schema:
            type: string
            format: uuid
            title: Session-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { }
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      summary: Получить пост
      operationId: get_post_post__post_id__get
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
            title: Post Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { }
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  '/posts/{user_login}':
    get:
      summary: Получить все посты по конкретному логину
      operationId: get_posts_posts__user_login__get
      parameters:
        - name: user_login
          in: path
          required: true
          schema:
            type: string
            title: User Login
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { }
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /like:
    post:
      summary: Поставить лайк посту
      operationId: like_post_like_post
      parameters:
        - name: post_id
          in: query
          required: true
          schema:
            type: integer
            title: Post Id
        - name: session-key
          in: header
          required: true
          schema:
            type: string
            format: uuid
            title: Session-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { }
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /view:
    post:
      summary: Просмотреть пост
      operationId: view_post_view_post
      parameters:
        - name: post_id
          in: query
          required: true
          schema:
            type: integer
            title: Post Id
        - name: session-key
          in: header
          required: true
          schema:
            type: string
            format: uuid
            title: Session-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { }
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /user:
    put:
      summary: Обновление информации о пользователе
      operationId: update_info_user_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Body_update_info_user_put'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { }
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Создать нового пользователя
      operationId: create_user_user_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterData'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { }
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /user_authentication:
    post:
      summary: Аутентификация в системе по логину и паролю
      operationId: authenticate_user_user_authentication_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterData'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionKey'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    Body_update_info_user_put:
      properties:
        user_data:
          $ref: '#/components/schemas/UserUpdate'
        session_key:
          $ref: '#/components/schemas/SessionKey'
      type: object
      required:
        - user_data
        - session_key
      title: Body_update_info_user_put
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    PostCreate:
      properties:
        content:
          type: string
          title: Content
      type: object
      required:
        - content
      title: PostCreate
    PostUpdate:
      properties:
        id:
          type: integer
          title: Id
        content:
          type: string
          title: Content
      type: object
      required:
        - id
        - content
      title: PostUpdate
    RegisterData:
      properties:
        login:
          type: string
          title: Login
        password:
          type: string
          title: Password
      type: object
      required:
        - login
        - password
      title: RegisterData
    SessionKey:
      properties:
        key:
          type: string
          format: uuid
          title: Key
      type: object
      required:
        - key
      title: SessionKey
    UserUpdate:
      properties:
        login:
          type: string
          title: Login
        name:
          anyOf:
            - type: string
            - type: 'null'
          title: Name
        surname:
          anyOf:
            - type: string
            - type: 'null'
          title: Surname
        birthdate:
          anyOf:
            - type: string
              format: date
            - type: 'null'
          title: Birthdate
        mail:
          anyOf:
            - type: string
            - type: 'null'
          title: Mail
        number_phone:
          anyOf:
            - type: string
            - type: 'null'
          title: Number Phone
      type: object
      required:
        - login
      title: UserUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError